selenide

+++ src/main/java/constant/Constants.java
@@ -2,7 +2,7 @@ package constant;
 
 public interface Constants {
      int IMPLICITLY_WAIT_TIMEOUT = 20;
-     int TIME_TO_WAIT_ELEMENT = 10;
+     long TIME_TO_WAIT_ELEMENT = 10;
      String INITIAL_PAGE_URL="https://www.bookdepository.com/";
 String INITIAL_PAGE_NAME="Initial home page";
 }
